About
"Chada Tech is a software engineering company that specializes in custom software design and development for both domestic and international clients. You read more about the company and decide this place would be a great fit for you. You are also happy to find that there is an opening for a junior developer and apply for the position immediately. The skills that you have acquired at SNHU have prepared you for this role. Consequently, you are selected for an interview!

As part of the Chada Tech interview process, you are asked to complete a coding skills test to “audition” for the role. This is a typical interviewing strategy that you may encounter in the field. Companies utilize this technique to ensure candidates can in fact do what they claim they can do. Therefore, this exercise will be based on a problem similar to what you will face in the junior developer role, and you will need to come up with a solution. This is your chance to really show them what you can do!" This project is my audition, for feedback I forgot to zip the project folder, and had some problems with with my menu and imcrenmenting time, but my code was well written and comments were used really well, according to my professor.